Move legacy annotations to org.javai.punit.api.legacy
The annotation namespace at org.javai.punit.api needs to be free for
the typed-compositional API's @ProbabilisticTest and @Experiment
annotations (Stage 5 — JUnit-driven entry point). The pre-Stage-3.5
annotations move to a sibling .legacy package: same names, different
package, no collision.

Mechanical rename + bulk import update across punit-core, punit-junit5,
and punit-sentinel. The legacy annotations themselves are untouched
in body — they keep working — but their location says "this is the
pre-typed authoring surface" plainly. Stage 8 retires them.
