+++ punit-core/src/main/java/org/javai/punit/ptest/engine/ProbabilisticTestValidator.java
@@ -2,7 +2,7 @@ package org.javai.punit.ptest.engine;
 
 import java.util.ArrayList;
 import java.util.List;
-import org.javai.punit.api.ProbabilisticTest;
+import org.javai.punit.api.legacy.ProbabilisticTest;
 import org.javai.punit.api.ThresholdOrigin;
 import org.javai.punit.spec.model.ExecutionSpecification;
