+++ punit-core/src/test/java/org/javai/punit/ptest/engine/ConfigurationResolverTest.java
@@ -7,7 +7,7 @@ import java.util.Map;
 import org.javai.punit.api.BudgetExhaustedBehavior;
 import org.javai.punit.api.ExceptionHandling;
 import org.javai.punit.api.Latency;
-import org.javai.punit.api.ProbabilisticTest;
+import org.javai.punit.api.legacy.ProbabilisticTest;
 import org.javai.punit.api.TestIntent;
 import org.javai.punit.api.ThresholdOrigin;
 import org.junit.jupiter.api.AfterEach;
