+++ punit-junit5/src/main/java/org/javai/punit/ptest/bernoulli/BernoulliTrialsStrategy.java
@@ -12,7 +12,7 @@ import org.javai.punit.api.BudgetExhaustedBehavior;
 import org.javai.punit.api.FactorSource;
 import org.javai.punit.api.HashableFactorSource;
 import org.javai.punit.api.InputSource;
-import org.javai.punit.api.ProbabilisticTest;
+import org.javai.punit.api.legacy.ProbabilisticTest;
 import org.javai.punit.api.TokenChargeRecorder;
 import org.javai.punit.controls.budget.BudgetOrchestrator;
 import org.javai.punit.controls.budget.CostBudgetMonitor;
