+++ punit-junit5/src/main/java/org/javai/punit/ptest/engine/OperationalApproachResolver.java
@@ -1,7 +1,7 @@
 package org.javai.punit.ptest.engine;
 
 import java.util.List;
-import org.javai.punit.api.ProbabilisticTest;
+import org.javai.punit.api.legacy.ProbabilisticTest;
 import org.javai.punit.spec.model.ExecutionSpecification;
 import org.javai.punit.statistics.OperationalApproach;
