+++ punit-junit5/src/main/java/org/javai/punit/ptest/engine/ProbabilisticTestExtension.java
@@ -10,7 +10,7 @@ import java.util.concurrent.atomic.AtomicInteger;
 import java.util.stream.Stream;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
-import org.javai.punit.api.ProbabilisticTest;
+import org.javai.punit.api.legacy.ProbabilisticTest;
 import org.javai.punit.api.TestIntent;
 import org.javai.punit.api.UseCaseProvider;
 import org.javai.punit.controls.budget.CostBudgetMonitor;
