+++ punit-junit5/src/main/java/org/javai/punit/ptest/strategy/ProbabilisticTestStrategy.java
@@ -2,7 +2,7 @@ package org.javai.punit.ptest.strategy;
 
 import java.lang.reflect.Method;
 import java.util.stream.Stream;
-import org.javai.punit.api.ProbabilisticTest;
+import org.javai.punit.api.legacy.ProbabilisticTest;
 import org.javai.punit.ptest.bernoulli.SampleResultAggregator;
 import org.javai.punit.ptest.engine.ConfigurationResolver;
 import org.junit.jupiter.api.extension.ExtensionContext;
