+++ punit-junit5/src/test/java/org/javai/punit/experiment/engine/ExperimentPacingIntegrationTest.java
@@ -7,11 +7,11 @@ import java.util.Map;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.concurrent.atomic.AtomicLong;
 import java.util.stream.Stream;
-import org.javai.punit.api.ExploreExperiment;
+import org.javai.punit.api.legacy.ExploreExperiment;
 import org.javai.punit.api.Factor;
 import org.javai.punit.api.FactorArguments;
 import org.javai.punit.api.FactorSource;
-import org.javai.punit.api.MeasureExperiment;
+import org.javai.punit.api.legacy.MeasureExperiment;
 import org.javai.punit.api.OutcomeCaptor;
 import org.javai.punit.api.Pacing;
 import org.javai.punit.api.UseCase;
