+++ punit-junit5/src/test/java/org/javai/punit/ptest/engine/ProbabilisticTestExtensionCovariateTest.java
@@ -8,7 +8,7 @@ import java.nio.file.Path;
 import java.security.MessageDigest;
 import java.security.NoSuchAlgorithmException;
 import java.util.HexFormat;
-import org.javai.punit.api.ProbabilisticTest;
+import org.javai.punit.api.legacy.ProbabilisticTest;
 import org.javai.punit.api.TestIntent;
 import org.javai.punit.api.UseCase;
 import org.javai.punit.model.CovariateProfile;
