+++ punit-junit5/src/test/java/org/javai/punit/testsubjects/InheritanceTestSubjects.java
@@ -7,9 +7,9 @@ import java.time.Instant;
 import java.util.Map;
 import java.util.stream.Stream;
 import org.javai.punit.api.InputSource;
-import org.javai.punit.api.MeasureExperiment;
+import org.javai.punit.api.legacy.MeasureExperiment;
 import org.javai.punit.api.OutcomeCaptor;
-import org.javai.punit.api.ProbabilisticTest;
+import org.javai.punit.api.legacy.ProbabilisticTest;
 import org.javai.punit.api.Sentinel;
 import org.javai.punit.api.TestIntent;
 import org.javai.punit.api.UseCaseProvider;
