+++ punit-sentinel/src/main/java/org/javai/punit/sentinel/SentinelClassIntrospector.java
@@ -5,8 +5,8 @@ import java.lang.reflect.Method;
 import java.util.ArrayList;
 import java.util.List;
 import org.javai.punit.api.InputSource;
-import org.javai.punit.api.MeasureExperiment;
-import org.javai.punit.api.ProbabilisticTest;
+import org.javai.punit.api.legacy.MeasureExperiment;
+import org.javai.punit.api.legacy.ProbabilisticTest;
 import org.javai.punit.api.Sentinel;
 import org.javai.punit.experiment.engine.input.InputSourceResolver;
 import org.javai.punit.usecase.UseCaseFactory;
