+++ punit-sentinel/src/main/java/org/javai/punit/sentinel/SentinelExperimentExecutor.java
@@ -10,7 +10,7 @@ import java.util.List;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.javai.punit.api.BudgetExhaustedBehavior;
-import org.javai.punit.api.MeasureExperiment;
+import org.javai.punit.api.legacy.MeasureExperiment;
 import org.javai.punit.api.OutcomeCaptor;
 import org.javai.punit.contract.UseCaseOutcome;
 import org.javai.punit.controls.budget.CostBudgetMonitor;
