+++ punit-sentinel/src/main/java/org/javai/punit/sentinel/SentinelRunner.java
@@ -8,8 +8,8 @@ import java.util.List;
 import java.util.Objects;
 import java.util.Optional;
 import java.util.function.Predicate;
-import org.javai.punit.api.MeasureExperiment;
-import org.javai.punit.api.ProbabilisticTest;
+import org.javai.punit.api.legacy.MeasureExperiment;
+import org.javai.punit.api.legacy.ProbabilisticTest;
 import org.javai.punit.ptest.bernoulli.FinalVerdictDecider;
 import org.javai.punit.ptest.engine.ConfigurationResolver;
 import org.javai.punit.statistics.ThresholdDeriver;
