+++ punit-sentinel/src/main/java/org/javai/punit/sentinel/SentinelTestExecutor.java
@@ -5,7 +5,7 @@ import java.util.List;
 import java.util.Optional;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
-import org.javai.punit.api.ProbabilisticTest;
+import org.javai.punit.api.legacy.ProbabilisticTest;
 import org.javai.punit.api.TestIntent;
 import org.javai.punit.controls.budget.CostBudgetMonitor;
 import org.javai.punit.model.TerminationReason;
